# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: 智慧停车应用
  version: 1.0.0
  title: 智慧停车应用
basePath: "/"
schemes:
- http

##认证
#securityDefinitions:
#  api_key:
#    type: "apiKey"
#    name: "token"
#    in: "header"

#定义接口数据
paths:
...
  /merchant/Advertisement/category/info/{id}:
    get:
      tags:
      - 营销管理
      summary: 获取广告类型详情
      produces:
      - application/json
      parameters:
      - name: id
        description: 广告类型id
        in: path
        type: String
        required: true
      responses:
        '200':
          description: 返回广告类型信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              data:
                description: 菜单信息
                $ref: '#/definitions/AdvertisementCategory'
  /merchant/Advertisement/category/add:
    post:
      tags:
      - 营销管理
      summary: 添加广告类型
      produces:
      - application/json
      parameters:
      - name: body
        description: 广告类型对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/AdvertisementCategorySave'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /merchant/Advertisement/category/update:
    patch:
      tags:
      - 营销管理
      summary: 修改广告类型
      produces:
      - application/json
      parameters:
      - name: body
        description: 广告类型对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/AdvertisementCategoryEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /merchant/Advertisement/category/delete/{id}:
    delete:
      tags:
      - 营销管理
      summary: 广告类型删除
      produces:
      - application/json
      parameters:
      - name: id
        description: 广告类型id
        in: path
        type: String
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  AdvertisementCategory:
    type: object
    properties:
      id:
        description: 广告类型id
        type: string
      tempCategoryCode:
        description: 广告类型代码
        type: string
      tempCategoryDesc:
        description: 广告类型名称
        type: string
      templateStatus:
        description: 广告类型状态
        type: integer
        format: int64
      createBy:
        description: 广告创建人
        type: integer
        format: int64
      createTime:
        description: 广告创建时间
        type: string
        format: date-time
      lastUpdatedBy:
        description: 广告最后更新人
        type: integer
        format: int64
      lastUpdatedTime:
        description: 广告最后更新时间
        type: string
        format: date-time
  AdvertisementCategorySave:
    type: object
    properties:
      tempCategoryDesc:
        description: 广告类型名称
        type: string
  AdvertisementCategoryEdit:
    type: object
    properties:
      id:
        description: 广告类型id
        type: string
      tempCategoryDesc:
        description: 广告类型名称
        type: string
      templateStatus:
        description: 广告类型状态
        type: integer
        format: int64
...